Refactor(anta.cli)!: Cleaner entrypoint
